pyproject.toml
[project]
name = "food-reco"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.6",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "dataclasses-json==0.6.7",
    "frozenlist==1.6.0",
    "greenlet==3.2.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.25",
    "langchain-community==0.3.24",
    "langchain-core==0.3.63",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.7",
    "langgraph-checkpoint==2.0.26",
    "langgraph-prebuilt==0.2.2",
    "langgraph-sdk==0.1.70",
    "langsmith==0.3.43",
    "marshmallow==3.26.1",
    "multidict==6.4.4",
    "mypy-extensions==1.1.0",
    "numpy==2.2.6",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "packaging==24.2",
    "propcache==0.3.1",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "tenacity==9.1.2",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "urllib3==2.4.0",
    "xxhash==3.5.0",
    "yarl==1.20.0",
    "zstandard==0.23.0", # 이전에 uv pip install로 추가했던 것
    "langchain-google-community==2.0.7",
    "langchain-openai>=0.3.19",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
